definitions moved in a implementation file
